+++ react/components/Loading.tsx
@@ -1,5 +1,8 @@
 import React, {CSSProperties, PureComponent} from 'react'
-import {getExtensionImplementation} from '../utils/assets'
+import {Instagram, List} from 'react-content-loader'
+
+import { TreePathContext, TreePathProps } from '../utils/treePath'
+
 import {RenderContextProps, withRuntimeContext} from './RenderContext'
 
 interface State {
@@ -11,20 +14,12 @@ const LOADING_TRESHOLD_MS = 1000
 const LOADING_UNMOUNT_TRESHOLD_MS = 1000
 let visible = false
 
-const defaultStyle = {
-  display: 'flex',
-  justifyContent: 'center',
-  padding: '50px',
-  width: '100%',
-}
-
 const defaultLoading = (
-  <div style={defaultStyle}>
-    <svg width="26px" height="26px" viewBox="0 0 100 100" preserveAspectRatio="xMidYMid">
-      <circle cx="50" opacity="0.4" cy="50" fill="none" stroke="#F71963" strokeWidth="14" r="40"></circle>
-      <circle cx="50" cy="50" fill="none" stroke="#F71963" strokeWidth="12" r="40" strokeDasharray="60 900" strokeLinecap="round" transform="rotate(96 50 50)">
-        <animateTransform attributeName="transform" type="rotate" calcMode="linear" values="0 50 50;360 50 50" keyTimes="0;1" dur="0.7s" begin="0s" repeatCount="indefinite"></animateTransform>
-      </circle>
+  <div className="flex justify-center w-100 pa4">
+    <svg width="26px" height="26px" viewBox="0 0 100 100" preserveAspectRatio="xMidYMid"
+          xmlns="http://www.w3.org/2000/svg"
+          xmlnsXlink="http://www.w3.org/1999/xlink">
+      <use xlinkHref="#sti-loading" />
     </svg>
   </div>
 )
@@ -55,15 +50,27 @@ class Loading extends PureComponent<RenderContextProps, State> {
   }
 
   public render() {
-    const {runtime: {extensions, page}} = this.props
-    const {visible: isVisible} = this.state
+    const { runtime: { extensions } } = this.props
+    const { visible: isVisible } = this.state
     const style: CSSProperties = { visibility: 'hidden' }
-    const [root] = page.split('/')
-    const LoadingExtension = getExtensionImplementation(extensions, `${root}/__loading`)
 
-    return <div style={isVisible ? undefined : style}>
-      { LoadingExtension ? <LoadingExtension /> : defaultLoading }
-    </div>
+    return (
+      <TreePathContext.Consumer>
+        {(value: TreePathProps) => {
+          const t = value.treePath
+          const loadingType = value.treePath && extensions[t] && extensions[t].preview && extensions[t].preview!.type
+          const loadingComponent = loadingType
+            ? loadingType === 'text' ? List : Instagram
+            : defaultLoading
+
+          return (
+            <div style={isVisible ? undefined : style}>
+              { loadingComponent }
+            </div>
+          )
+        }}
+      </TreePathContext.Consumer>
+    )
   }
 }