+++ react/utils/routes.ts
@@ -1,5 +1,6 @@
-import defaultPage from './defaultPage.graphql'
-import pageQuery from './Page.graphql'
+import navigationPageQuery from '../queries/navigationPage.graphql'
+import routePreviews from '../queries/routePreviews.graphql'
+import { parseMessages } from './messages'
 
 const parsePageQueryResponse = (page: PageQueryResponse): ParsedPageQueryResponse => {
   const {
@@ -8,15 +9,14 @@ const parsePageQueryResponse = (page: PageQueryResponse): ParsedPageQueryRespons
     cacheHintsJSON,
     componentsJSON,
     extensionsJSON,
-    messagesJSON,
+    messages,
     pagesJSON,
   } = page
 
-  const [cacheHints, components, extensions, messages, pages, settings] = [
+  const [cacheHints, components, extensions, pages, settings] = [
     cacheHintsJSON,
     componentsJSON,
     extensionsJSON,
-    messagesJSON,
     pagesJSON,
     appsSettingsJSON,
   ].map(json => JSON.parse(json))
@@ -26,7 +26,7 @@ const parsePageQueryResponse = (page: PageQueryResponse): ParsedPageQueryRespons
     cacheHints,
     components,
     extensions,
-    messages,
+    messages: parseMessages(messages),
     pages,
     settings
   }
@@ -36,61 +36,65 @@ const parseDefaultPagesQueryResponse = (defaultPages: DefaultPagesQueryResponse)
   const {
     componentsJSON,
     extensionsJSON,
-    messagesJSON,
+    messages,
   } = defaultPages
 
-  const [components, extensions, messages] = [
+  const [components, extensions] = [
     componentsJSON,
     extensionsJSON,
-    messagesJSON,
   ].map(json => JSON.parse(json))
 
   return {
     components,
     extensions,
-    messages,
+    messages: parseMessages(messages),
   }
 }
 
-export const fetchRoutes = ({
+export const fetchNavigationPage = ({
   apolloClient,
-  conditions,
-  device,
-  locale,
-  page,
-  params,
-  path,
+  routeId,
+  declarer,
   production,
-  renderMajor,
-  scope,
-  template,
-}: FetchRoutesInput) => apolloClient.query<{page: PageQueryResponse}>({
+  paramsJSON,
+  renderMajor
+}: FetchNavigationDataInput) => apolloClient.query<{navigationPage: PageQueryResponse}>({
   fetchPolicy: production ? 'cache-first' : 'network-only',
-  query: pageQuery,
+  query: navigationPageQuery,
   variables: {
-    conditions,
-    device,
-    locale,
-    page,
-    params,
-    path,
+    declarer,
+    params: paramsJSON,
     production,
     renderMajor,
-    scope,
-    template,
+    routeId,
   }
-}).then<ParsedPageQueryResponse>(({data: {page: pageData}, errors}: PageQueryResult) =>
+}).then<ParsedPageQueryResponse>(({data: {navigationPage: pageData}, errors}: GraphQLResult<'navigationPage', PageQueryResponse>) =>
       errors ? Promise.reject(errors) : parsePageQueryResponse(pageData))
+
+const getRoutesParam = (routeIds: string[], pages: Pages) => {
+  return routeIds
+    .filter(routeId => routeId in pages)
+    .map(routeId => {
+      const page = pages[routeId]
+      return {
+        declarer: page.declarer,
+        routeId,
+      }
+    })
+}
 
 export const fetchDefaultPages = ({
   apolloClient,
-  locale,
+  pages,
   routeIds,
-}: FetchDefaultPages) => apolloClient.query<{defaultPages: DefaultPagesQueryResponse}>({
-  query: defaultPage,
-  variables: {
-    locale,
-    routeIds
-  }
-}).then<ParsedDefaultPagesQueryResponse>(({data: {defaultPages}, errors}: DefaultPagesQueryResult) =>
-  errors ? Promise.reject(errors) : parseDefaultPagesQueryResponse(defaultPages))
+  renderMajor,
+}: FetchDefaultPages) => {
+  return apolloClient.query<{defaultPages: DefaultPagesQueryResponse}>({
+    query: routePreviews,
+    variables: { renderMajor, routes: getRoutesParam(routeIds, pages)}
+  }).then<ParsedDefaultPagesQueryResponse>(
+    ({data: {defaultPages}, errors}: DefaultPagesQueryResult) => {
+      return errors ? Promise.reject(errors) : parseDefaultPagesQueryResponse(defaultPages)
+    }
+  )
+}